docker, games, game template

+++ Admin/OnAim.Admin.Contracts/Dtos/Game/PriceDto.cs
@@ -4,6 +4,6 @@ public class PriceDto
 {
     public string Id { get; set; }
     public int Value { get; set; }
-    public int Multiplier { get; set; }
+    public decimal Multiplier { get; set; }
     public string CoinId { get; set; }
 }
